JAMES-2151 Move InMemorySieveQuotaRepository in a more generic module
